requirements.txt
openai
langchain
streamlit==1.24.0
python-dotenv
PyPDF2

-e .